refactor(pipeline): extract recovery functions to recovery.go [#466]
Move emitPhaseFailed and recoverCrashedPhases from engine.go to recovery.go.
